delete should work for up to 25 events per user, but going to add pagination to increase

+++ main.py
@@ -1,5 +1,5 @@
 import discord
-from ui import ScheduleModal
+from ui import DeleteDropdown, ScheduleModal
 from utilities import is_in_dms, get_days_until
 from discord import app_commands
 from discord.ext import commands
@@ -82,22 +82,27 @@ async def countdown(interaction: discord.Interaction) -> None:
 
 # --------Delete command--------
 @bot.tree.command(name="delete", description="delete an event")
-async def delete(interaction: discord.Interaction, name: str) -> None:
+async def delete(interaction: discord.Interaction) -> None:
     if is_in_dms(interaction):
         await interaction.response.send_message("This command only works in a server.")
         return
     sender = interaction.user.id
     guild_id = interaction.guild.id
     cur = con.cursor()
+    cur.execute("SELECT name FROM events WHERE guild_id = ? AND user_id = ?", (guild_id, sender))
+    eventname = await interaction.response.send_dropdown(DeleteDropdown([discord.SelectOption(label=row["name"]) for row in cur.fetchall()]))
+    if not eventname:
+        await interaction.response.send_message("No event selected.", ephemeral=True)
+        return
     cur.execute(
         "DELETE FROM events WHERE guild_id = ? AND user_id = ? AND name = ?",
-        (guild_id, sender, name),
+        (guild_id, sender, eventname),
     )
     if cur.rowcount == 0:
         await interaction.response.send_message("No event with that name found for you in this server")
         return
     con.commit()
-    await interaction.response.send_message(f"Deleted {name}")
+    await interaction.response.send_message(f"Deleted {eventname}")
 
 
 load_dotenv()
